feat(refactor): AS-963 / PR-05a-h4-c — final internal/session + internal/aws import elimination from internal/tui
The 5a-extract Stage 2 exit gate: docs/refactor/05-boundary.md:230-264 passes
for the first time. internal/tui's production import set no longer references
internal/aws or internal/session — only internal/domain, internal/runtime,
internal/resource, internal/config, internal/cache, internal/tui/{keys,layout,
styles,views,text}, and Bubble Tea / Lipgloss / textinput.

Files added:
  - internal/runtime/transport.go  — type ServiceClients = awsclient.ServiceClients
  - internal/runtime/relatedcache.go — RelatedCacheKey, RelatedCacheReplay free
    funcs + RelatedCacheResult alias so the renderer never imports internal/session
  - internal/runtime/accessors.go — typed Core accessors (Profile, Region,
    ConnectGen, Identity returns *domain.CallerIdentity, ResourceCache returns
    *domain.ListViewCacheEntry, LazyResourceCache, SetIdentityFetching,
    ClearCommand, HasIssueEnricher) + Bootstrap convenience constructor that
    internalises session.New() so tui.New no longer imports internal/session
  - internal/domain/resource_cache.go — ListViewCacheEntry value type moved
    from internal/session (renamed from ResourceCacheEntry to avoid the
    name collision with the existing domain.ResourceCacheEntry related-checker
    snapshot type; session.ResourceCacheEntry stays available as a type alias)

Files modified:
  - internal/session/session.go — ResourceCacheEntry becomes an alias to
    domain.ListViewCacheEntry; all existing callers compile unchanged
  - internal/tui/app.go — dropped awsclient + session imports; WithClients
    signature now *runtime.ServiceClients; cacheTopLevelResourceList literal
    now &domain.ListViewCacheEntry{}; session.New() replaced by
    runtime.Bootstrap(profile, region, types). Shrunk from 668 → 308 LOC by
    splitting view-shell helpers into app_view.go, option setters into
    app_options.go, and view-stack helpers into app_stack.go (within spec
    300-400 budget).
  - internal/tui/app_accessors.go — deleted the Session() *session.Session
    accessor that was the last production-side leak; added Core() *runtime.Core
    for test ergonomics. Test sites migrated from m.Session() to
    m.Core().Session() across 14 test files / 136 call sites.
  - internal/tui/app_dispatch.go — session.RelatedCacheKey / RelatedCacheResult
    references switched to runtime.* equivalents
  - internal/tui/probe_adapter.go — awsclient.Wave2EnricherFor replaced by
    m.core.HasIssueEnricher
  - internal/tui/runtime_adapter_navigate.go — same enricher swap;
    session.RelatedCacheKey / Replay switched to runtime.*
  - internal/tui/runtime_adapter_related.go — session.RelatedCacheKey / Replay
    switched to runtime.*
  - internal/tui/runtime_adapter.go — historical comment scrub
  - internal/tui/app_options.go, app_view.go, app_stack.go — extracted from
    app.go to honour the 300-400 LOC budget

Spec acceptance (every gate verified locally):

  go list -f '{{.Imports}}' github.com/k2m30/a9s/v3/internal/tui \
    | tr ' ' '\n' | grep -E 'internal/session|internal/aws$'
  # zero hits

  wc -l internal/tui/app.go
  # 308

  rg 'func \(m Model\) Session\(\)' internal/tui/
  # zero hits

  rg 'awsclient\.' internal/tui/ --type go | grep -v _test
  # zero hits

  rg '\bsession\.' internal/tui/ --type go | grep -v _test
  # zero hits

  go list -f '{{.Imports}}' github.com/k2m30/a9s/v3/internal/runtime \
    | grep -E 'bubbletea|lipgloss|bubbles'
  # zero hits

Pre-push gate sub-targets run locally (test-race deferred to CI — cgo
unavailable in this sandbox; gcc not installed): build, lint, security
(govulncheck), gofix, verify-readonly, verify-zero-init, check-readme,
snapshot, mdlint — all PASS. `go test ./...` PASS.

Spec: docs/refactor/05-pr-05a-h4.md:496-611.
Parent: AS-960. Predecessor: AS-962 (h4-b, PR #411).
