+++ internal/tui/app.go
@@ -3,23 +3,17 @@ package tui
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"charm.land/bubbles/v2/textinput"
 	tea "charm.land/bubbletea/v2"
-	"charm.land/lipgloss/v2"
 
-	awsclient "github.com/k2m30/a9s/v3/internal/aws"
 	"github.com/k2m30/a9s/v3/internal/config"
 	"github.com/k2m30/a9s/v3/internal/domain"
 	"github.com/k2m30/a9s/v3/internal/resource"
 	"github.com/k2m30/a9s/v3/internal/runtime"
-	"github.com/k2m30/a9s/v3/internal/session"
-	"github.com/k2m30/a9s/v3/internal/tui/keys"
-	"github.com/k2m30/a9s/v3/internal/tui/layout"
 	"github.com/k2m30/a9s/v3/internal/runtime/messages"
-	"github.com/k2m30/a9s/v3/internal/tui/styles"
+	"github.com/k2m30/a9s/v3/internal/tui/keys"
 	"github.com/k2m30/a9s/v3/internal/tui/views"
 )
 
@@ -53,8 +47,8 @@ type errorEntry struct {
 // AWS clients, and routes all messages to the active child view.
 //
 // Session state is owned exclusively by core via core.Session(). Call
-// m.core.Session() at each use site. See internal/session/session.go for the
-// ownership contract. handleProfileSelected / handleRegionSelected call
+// m.core.Session() at each use site. See the internal/session package for
+// the ownership contract. handleProfileSelected / handleRegionSelected call
 // m.core.Session().Rotate() to invalidate in-flight async results on switch.
 type Model struct {
 	core    *runtime.Core    // platform-agnostic app core
@@ -107,56 +101,6 @@ type Model struct {
 // Option configures the root Model.
 type Option func(*Model)
 
-// WithProfile overrides the profile field on the session. Used in tests that need
-// a specific profile string without going through the live AWS bootstrap path.
-func WithProfile(profile string) Option {
-	return func(m *Model) { m.core.Session().Profile = profile }
-}
-
-// WithRegion overrides the region field on the session. Used in tests that need
-// a specific region string without going through the live AWS bootstrap path.
-func WithRegion(region string) Option {
-	return func(m *Model) { m.core.Session().Region = region }
-}
-
-// WithIsDemo marks the session as demo mode, which skips Wave 2 enrichment.
-// Set by the --demo CLI bootstrap path. Distinct from WithNoCache which only
-// disables disk persistence.
-func WithIsDemo(demo bool) Option {
-	return func(m *Model) {
-		m.isDemo = demo
-	}
-}
-
-// WithNoCache disables resource availability caching and background checks.
-func WithNoCache(disabled bool) Option {
-	return func(m *Model) {
-		m.core.Session().NoCache = disabled
-	}
-}
-
-// WithClients pre-supplies a set of service clients so that Init() emits a
-// synthetic ClientsReadyMsg instead of initiating a live AWS connection.
-func WithClients(clients *awsclient.ServiceClients) Option {
-	return func(m *Model) {
-		m.core.Session().PreSuppliedClients = clients
-	}
-}
-
-// WithActiveTheme sets the initial active theme filename for the selector's
-// "(current)" indicator. main.go passes the validated theme after loading it.
-func WithActiveTheme(name string) Option {
-	return func(m *Model) { m.activeTheme = name }
-}
-
-// WithCommand sets the initial resource short name to navigate to on the first
-// ClientsReadyMsg. Used by the -c/--command CLI flag to open a resource list
-// directly on startup instead of the main menu. The caller is responsible for
-// resolving the input via resource.FindResourceType.
-func WithCommand(name string) Option {
-	return func(m *Model) { m.core.Session().Command = name }
-}
-
 // New constructs the initial Model.
 func New(profile, region string, opts ...Option) Model {
 	ti := textinput.New()
@@ -175,11 +119,8 @@ func New(profile, region string, opts ...Option) Model {
 	// construction and threaded through all fetchers.
 	ctx, cancel := context.WithCancel(context.Background())
 
-	sess := session.New()
-	sess.Profile = profile
-	sess.Region = region
 	m := Model{
-		core:        runtime.New(sess, resource.AllResourceTypes()),
+		core:        runtime.Bootstrap(profile, region, resource.AllResourceTypes()),
 		keys:        k,
 		stack:       []views.View{&menu},
 		cmdInput:    ti,
@@ -209,7 +150,7 @@ func (m Model) Cancel() {
 	}
 }
 
-// Init implements tea.Model. Fires a command to establish the AWS session.
+// Init implements tea.Model. Fires a command to establish the live AWS connection.
 // When pre-supplied clients are present (demo mode or tests), emits a synthetic
 // ClientsReadyMsg immediately. Otherwise initiates the live AWS connection flow.
 func (m Model) Init() tea.Cmd {
@@ -365,304 +306,3 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return m.updateActiveView(msg)
 }
 
-// View implements tea.Model. Composes header + frame around active view content.
-func (m Model) View() tea.View {
-	alt := func(s string) tea.View {
-		v := tea.NewView(s)
-		v.AltScreen = true
-		return v
-	}
-	if m.width == 0 {
-		return alt("")
-	}
-	if m.width < layout.MinTerminalWidth {
-		return alt("Terminal too narrow (min 60 columns)")
-	}
-	if m.height < 7 {
-		return alt("Terminal too short (min 7 lines)")
-	}
-
-	active := m.activeView()
-
-	headerProfile := m.core.Session().Profile
-	headerRegion := m.core.Session().Region
-	if sel, ok := active.(*views.SelectorModel); ok {
-		if sel.Title() == "aws-regions" {
-			headerRegion = "..."
-		} else if sel.Title() == "aws-profiles" {
-			headerProfile = "..."
-		}
-	}
-	rightContent := m.headerRight()
-	badge := m.accountBadge()
-	role := m.identityRoleName()
-	// §7.4: when stack depth exceeds 4, show "[N]" in place of the version string.
-	displayVersion := Version
-	if len(m.stack) > 4 {
-		displayVersion = fmt.Sprintf("[%d]", len(m.stack))
-	}
-	cacheKey := headerProfile + ":" + headerRegion + ":" + displayVersion + ":" + rightContent + ":" + badge + ":" + role + ":" + fmt.Sprintf("%d", m.width)
-	header := m.headerCache
-	if cacheKey != m.headerCacheKey {
-		header = layout.RenderHeader(headerProfile, headerRegion, displayVersion, m.width, rightContent, badge, role)
-		m.headerCache = header
-		m.headerCacheKey = cacheKey
-	}
-
-	content := active.View()
-	var lines []string
-	if content != "" {
-		lines = strings.Split(content, "\n")
-	}
-	frameHeight := max(m.height-1, 3)
-	frameTitle := active.FrameTitle()
-	if d, ok := active.(*views.DetailModel); ok {
-		src := d.SourceResource()
-		if src.ID != "" {
-			if src.Name != "" {
-				frameTitle = fmt.Sprintf("detail -- %s (%s)", src.ID, src.Name)
-			} else {
-				frameTitle = "detail -- " + src.ID
-			}
-		}
-	}
-	var hints []layout.KeyHint
-	if h, ok := active.(views.Hintable); ok {
-		hints = h.BottomHints()
-	}
-	frame := layout.RenderFrameWithHints(lines, frameTitle, hints, m.width, frameHeight)
-
-	v := tea.NewView(header + "\n" + frame)
-	v.AltScreen = true
-	return v
-}
-
-// activeView returns the top of the view stack.
-func (m *Model) activeView() views.View {
-	return m.stack[len(m.stack)-1]
-}
-
-// pushView adds a new view to the stack.
-func (m *Model) pushView(v views.View) {
-	m.stack = append(m.stack, v)
-}
-
-// popView removes the top view. Returns false if only one entry remains.
-func (m *Model) popView() bool {
-	if len(m.stack) <= 1 {
-		return false
-	}
-	// Sync list count back to main menu when popping directly from list → menu.
-	// Only depth 2 (menu → list) triggers this. Related lists (pushed from a detail
-	// view at depth 3+) are marked escPops=true because they show filtered subsets
-	// of a resource type, not the global population — syncing their count back to
-	// the menu badge would overwrite the real global count with a filter result.
-	// See app_related.go for related-list construction.
-	if len(m.stack) == 2 {
-		if rl, ok := m.stack[1].(*views.ResourceListModel); ok {
-			if menu, ok := m.stack[0].(*views.MainMenuModel); ok {
-				shortName := rl.ShortName()
-				if shortName != "" {
-					newCount := rl.LoadedCount()
-					newTrunc := rl.IsTruncated()
-					curCount, known := menu.GetAvailability()[shortName]
-					if !newTrunc || !known || newCount > curCount {
-						menu.SetAvailability(shortName, newCount)
-						menu.SetTruncated(shortName, newTrunc)
-					}
-					// Sync-back issue count with only-increase guard (T036, FR-022).
-					// The list's Status-based issueCount may be lower than the menu's
-					// enriched count. Never overwrite a higher enriched count.
-					newIssues := rl.IssueCount()
-					curIssues := menu.GetIssueCounts()[shortName]
-					curIssueTrunc := menu.GetIssueTruncated()[shortName]
-					switch {
-					case newIssues > curIssues:
-						// higher enriched count from rl: take it
-						menu.SetIssues(shortName, newIssues, newTrunc)
-					case newIssues == curIssues && curIssueTrunc && !newTrunc:
-						// count confirmed but stale "+" must clear
-						menu.SetIssues(shortName, newIssues, false)
-					}
-				}
-			}
-		}
-	}
-	m.stack = m.stack[:len(m.stack)-1]
-	return true
-}
-
-// innerSize returns the content area dimensions inside the frame.
-func (m *Model) innerSize() (int, int) {
-	w := m.width - 2
-	h := m.height - 3
-	if w < 1 {
-		w = 1
-	}
-	if h < 1 {
-		h = 1
-	}
-	return w, h
-}
-
-// propagateSize calls SetSize on every view in the stack with inner dimensions.
-func (m *Model) propagateSize() {
-	w, h := m.innerSize()
-	for _, v := range m.stack {
-		v.SetSize(w, h)
-	}
-}
-
-// updateActiveView delegates a message to the active view and merges the result.
-func (m Model) updateActiveView(msg tea.Msg) (tea.Model, tea.Cmd) {
-	active := m.activeView()
-	switch v := active.(type) {
-	case *views.MainMenuModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.ResourceListModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		m.cacheTopLevelResourceList(updated)
-		return m, cmd
-	case *views.DetailModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.YAMLModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.JSONModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.RevealModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.SelectorModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.HelpModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	case *views.IdentityModel:
-		updated, cmd := v.Update(msg)
-		m.stack[len(m.stack)-1] = &updated
-		return m, cmd
-	}
-	return m, nil
-}
-
-func (m *Model) cacheTopLevelResourceList(rl views.ResourceListModel) {
-	if rl.ParentContext() != nil || rl.EscPops() {
-		return
-	}
-	rt := rl.ResourceType()
-	sortColIdx, sortAsc := rl.SortState()
-	m.core.Session().ResourceCache[rt] = &session.ResourceCacheEntry{
-		Resources:     rl.AllResources(),
-		Pagination:    rl.PaginationState(),
-		FilterText:    rl.FilterText(),
-		AttentionOnly: rl.AttentionOnly(),
-		SortColIdx:    sortColIdx,
-		SortAsc:       sortAsc,
-		CursorPos:     rl.CursorPosition(),
-		HScrollOffset: rl.HScrollOffset(),
-	}
-}
-
-// helpContext determines the HelpContext from the current active view.
-func (m *Model) helpContext() views.HelpContext {
-	return m.activeView().GetHelpContext()
-}
-
-// headerRight returns the pre-rendered right-side string for the header.
-func (m Model) headerRight() string {
-	switch m.inputMode {
-	case modeFilter:
-		return styles.FilterActive.Render("/" + m.cmdInput.Value())
-	case modeCommand:
-		return styles.FilterActive.Render(":" + m.cmdInput.Value())
-	}
-	// Show search state from active view.
-	if s, ok := m.activeView().(views.Searchable); ok && s.IsSearchActive() {
-		info := s.SearchInfo()
-		if info != "" {
-			return styles.FilterActive.Render(info)
-		}
-	}
-	if m.flash.active {
-		text := m.flash.text
-		// Truncate to prevent header wrapping (fixes #84).
-		// Reserve ~40 chars for the left side (a9s + version + profile:region + padding).
-		// Errors get more header width; reserve 6 chars minimum for the brand + gap.
-		maxFlash := max(m.width-40, 20)
-		if m.flash.isError {
-			maxFlash = max(m.width-6, 20) // errors get wider display; 6 = innerPad(2)+minLeft(3)+gap(1)
-		}
-		if lipgloss.Width(text) > maxFlash {
-			// Truncate by runes to handle Unicode safely.
-			runes := []rune(text)
-			if len(runes) > maxFlash-3 {
-				text = string(runes[:maxFlash-3]) + "..."
-			}
-		}
-		if m.flash.isError {
-			return styles.FlashError.Render(text)
-		}
-		return styles.FlashSuccess.Render(text)
-	}
-	if rv, ok := m.activeView().(*views.RevealModel); ok {
-		return rv.HeaderWarning()
-	}
-	if m.showErrorHint && len(m.errorHistory) > 0 {
-		return styles.FlashError.Render("! for errors")
-	}
-	return styles.DimText.Render("? for help")
-}
-
-// accountBadge returns the account alias (preferred) or account ID for the header.
-func (m Model) accountBadge() string {
-	if m.core.Session().Identity == nil {
-		return ""
-	}
-	if m.core.Session().Identity.AccountAlias != "" {
-		return m.core.Session().Identity.AccountAlias
-	}
-	return m.core.Session().Identity.AccountID
-}
-
-// identityRoleName returns the identity name (role or user) for the header.
-func (m Model) identityRoleName() string {
-	if m.core.Session().Identity == nil {
-		return ""
-	}
-	return m.core.Session().Identity.IdentityName
-}
-
-// identityToViewData converts the session-cached awsclient.CallerIdentity
-// to a view-layer IdentityData. Used at IdentityModel construction time
-// (the `i` key press) to seed the view from current session state
-// before the in-flight identity fetch returns. The post-h4-b
-// SetIdentityIntent updates the IdentityModel via applyIntents using
-// the *domain.CallerIdentity mirror — this helper covers the construction
-// path that runs before any intent fires.
-func (m Model) identityToViewData() views.IdentityData {
-	if m.core.Session().Identity == nil {
-		return views.IdentityData{}
-	}
-	return views.IdentityData{
-		AccountID:     m.core.Session().Identity.AccountID,
-		AccountAlias:  m.core.Session().Identity.AccountAlias,
-		ARN:           m.core.Session().Identity.Arn,
-		RoleName:      m.core.Session().Identity.RoleName,
-		UserName:      m.core.Session().Identity.UserName,
-		SessionName:   m.core.Session().Identity.SessionName,
-		IsAssumedRole: m.core.Session().Identity.IsAssumedRole,
-	}
-}
