+++ internal/tui/runtime_adapter.go
@@ -177,7 +177,7 @@ func (m Model) runtimeTasksToCmd(tasks []runtime.TaskRequest) tea.Cmd {
 // detail enricher and emits an EnrichDetailResultMsg. It reads every
 // runtime-side input (DetailCtx, Generation) from the typed payload —
 // PR-05a-h4-b (AS-962) moved DetailEnrichmentCtx construction onto Core
-// so the adapter no longer touches awsclient.DetailEnrichmentCtx
+// so the adapter no longer touches the AWS-side DetailEnrichmentCtx
 // directly here. The remaining adapter-owned input is m.appCtx (the
 // app-wide cancellation context); ctx still wraps a 10 s per-call
 // timeout the runtime cannot express because tea.Cmd composition
