+++ tests/unit/qa_partial_success_handlers_test.go
@@ -67,7 +67,7 @@ func TestHandleAvailabilityChecked_PartialErrAppliesState(t *testing.T) {
 		Resources:    partialResources,
 		Issues:       0,
 		Truncated:    false,
-		Gen:          m.Session().AvailabilityGen,
+		Gen:          m.Core().Session().AvailabilityGen,
 	})
 
 	// CONTRACT 1: A FlashMsg with IsError=true must be emitted.
@@ -100,7 +100,7 @@ func TestHandleAvailabilityChecked_PartialErrAppliesState(t *testing.T) {
 	// retained, buildEnrichQueue includes "ec2" → enrichment is dispatched.
 	_, enrichCmd := rootApplyMsg(m, messages.AvailabilityChecked{
 		ResourceType: "dummy-for-finalize",
-		Gen:          m.Session().AvailabilityGen,
+		Gen:          m.Core().Session().AvailabilityGen,
 		Count:        0,
 		HasResources: false,
 	})
@@ -139,13 +139,13 @@ func TestHandleAvailabilityChecked_HardErr_NoStateApplied(t *testing.T) {
 		HasResources: false,
 		Count:        0,
 		Resources:    nil,
-		Gen:          m.Session().AvailabilityGen,
+		Gen:          m.Core().Session().AvailabilityGen,
 	})
 
 	// Hard failure: wave 2 must NOT be dispatched for lambda (no probe resources).
 	_, enrichCmd := rootApplyMsg(m, messages.AvailabilityChecked{
 		ResourceType: "dummy-finalize",
-		Gen:          m.Session().AvailabilityGen,
+		Gen:          m.Core().Session().AvailabilityGen,
 	})
 	if enrichCmd != nil {
 		enrichMsgs := collectEnrichmentMsgs(enrichCmd)
@@ -188,7 +188,7 @@ func TestHandleEnrichmentChecked_PartialErrAppliesState(t *testing.T) {
 	// Seed probeResources["ec2"] so the handler can merge FieldUpdates.
 	m, _ = rootApplyMsg(m, messages.AvailabilityChecked{
 		ResourceType: "ec2",
-		Gen:          m.Session().AvailabilityGen,
+		Gen:          m.Core().Session().AvailabilityGen,
 		Count:        1,
 		HasResources: true,
 		Resources: []resource.Resource{
