+++ tests/unit/qa_probe_truncation_per_type_test.go
@@ -96,7 +96,7 @@ func TestBuildResourceCacheSnapshot_ProbeAuthoritative_SinglePageComplete(t *tes
 		Resources:      map[string][]resource.Resource{targetType: {probeResource}},
 		// Stamp the live AvailabilityGen so the AS-657/AS-659 staleness guard
 		// accepts the message (AcceptZeroGen=false after AS-659).
-		Gen: m.Session().AvailabilityGen,
+		Gen: m.Core().Session().AvailabilityGen,
 	})
 
 	// Navigate to src detail view so RelatedCheckStartedMsg is handled.
@@ -200,7 +200,7 @@ func TestBuildResourceCacheSnapshot_ProbeTruncated_StampsTrue(t *testing.T) {
 		Resources:      map[string][]resource.Resource{targetType: {probeResource}},
 		// Stamp the live AvailabilityGen so the AS-657/AS-659 staleness guard
 		// accepts the message (AcceptZeroGen=false after AS-659).
-		Gen: m.Session().AvailabilityGen,
+		Gen: m.Core().Session().AvailabilityGen,
 	})
 
 	srcRes := resource.Resource{ID: "pt2-src-001", Name: "pt2-src-001"}
